# Changes   o Added Input system   o Updated Application, Layer and Utility classes   o Updated Window and Event system   o Updated Build System

+++ Source/Omnia/Debug/Memory.h
@@ -1,10 +1,10 @@
 #pragma once
 
 struct AllocationMetrics {
-	uint32_t TotalAllocated = 0;
-	uint32_t TotalFreed = 0;
+	size_t TotalAllocated = 0;
+	size_t TotalFreed = 0;
 
-	uint32_t CurrentUsage() { return TotalAllocated - TotalFreed; }
+	size_t CurrentUsage() { return TotalAllocated - TotalFreed; }
 };
 
 static AllocationMetrics iAllocationMetrics;
